feat: save sidebar state in local storage
